refactor: refact on category page
