+++ packages/theme/pages/Category.vue
@@ -4,94 +4,8 @@
       class="breadcrumbs desktop-only"
       :breadcrumbs="breadcrumbs"
     />
-    <div class="navbar section">
-      <div class="navbar__aside desktop-only">
-        <LazyHydrate never>
-          <SfHeading
-            :level="3"
-            :title="$t('Categories')"
-            class="navbar__title"
-          />
-        </LazyHydrate>
-      </div>
 
-      <div class="navbar__main">
-        <LazyHydrate on-interaction>
-          <SfButton
-            class="sf-button--text navbar__filters-button"
-            data-cy="category-btn_filters"
-            aria-label="Filters"
-            @click="toggleFilterSidebar"
-          >
-            <SfIcon
-              size="24px"
-              color="dark-secondary"
-              icon="filter2"
-              class="navbar__filters-icon"
-              data-cy="category-icon_"
-            />
-            {{ $t('Filters') }}
-          </SfButton>
-        </LazyHydrate>
-
-        <div class="navbar__sort desktop-only">
-          <span class="navbar__label">{{ $t('Sort by') }}:</span>
-          <LazyHydrate on-interaction>
-            <SfSelect
-              :value="sortBy.selected"
-              placeholder="Select sorting"
-              data-cy="category-select_sortBy"
-              class="navbar__select"
-              @input="th.changeSorting"
-            >
-              <SfSelectOption
-                v-for="(option, index) in sortBy.options"
-                :key="index"
-                :value="option.value"
-                class="sort-by__option"
-                >{{ option.attrName }}</SfSelectOption
-              >
-            </SfSelect>
-          </LazyHydrate>
-        </div>
-
-        <div class="navbar__counter">
-          <span class="navbar__label desktop-only"
-            >{{ $t('Products found') }}:
-          </span>
-          <span class="desktop-only">{{ pagination.totalItems }}</span>
-          <span class="navbar__label smartphone-only"
-            >{{ pagination.totalItems }} {{ $t('Items') }}</span
-          >
-        </div>
-
-        <div class="navbar__view">
-          <span class="navbar__view-label desktop-only">{{ $t('View') }}</span>
-          <SfIcon
-            data-cy="category-icon_grid-view"
-            class="navbar__view-icon"
-            :color="isCategoryGridView ? 'green' : 'dark-secondary'"
-            icon="tiles"
-            size="14px"
-            role="button"
-            aria-label="Change to grid view"
-            :aria-pressed="isCategoryGridView"
-            @click="changeToCategoryGridView"
-          />
-          <SfIcon
-            data-cy="category-icon_list-view"
-            class="navbar__view-icon"
-            :color="!isCategoryGridView ? 'green' : 'dark-secondary'"
-            icon="list"
-            size="14px"
-            role="button"
-            aria-label="Change to list view"
-            :aria-pressed="!isCategoryGridView"
-            @click="changeToCategoryListView"
-          />
-        </div>
-      </div>
-    </div>
+    <CategoryNavbar :facetsList="result" />
 
     <div class="main section">
       <div class="sidebar desktop-only">
@@ -320,107 +234,25 @@
     </div>
 
     <LazyHydrate when-idle>
-      <SfSidebar
-        :visible="isFilterSidebarOpen"
-        title="Filters"
-        class="sidebar-filters"
-        @close="toggleFilterSidebar"
-      >
-        <div class="filters desktop-only">
-          <div v-for="(facet, i) in facets" :key="i">
-            <template v-if="facetHasMoreThanOneOption(facet)">
-              <SfHeading
-                :level="4"
-                :title="facet.label"
-                class="filters__title sf-heading--left"
-                :key="`filter-title-${facet.value}`"
-              />
-              <div
-                v-if="isFacetColor(facet)"
-                class="filters__colors"
-                :key="`${facet.value}-colors`"
-              >
-                <SfColor
-                  v-for="option in facet.options"
-                  :key="`${facet.id}-${option.value}`"
-                  :data-cy="`category-filter_color_${option.value}`"
-                  :color="option.value"
-                  :selected="isFilterSelected(facet, option)"
-                  class="filters__color"
-                  @click="() => selectFilter(facet, option)"
-                />
-              </div>
-              <div v-else>
-                <SfFilter
-                  v-for="option in facet.options"
-                  :key="`${facet.id}-${option.value}`"
-                  :data-cy="`category-filter_${facet.id}_${option.value}`"
-                  :label="
-                    option.label + `${option.count ? ` (${option.count})` : ''}`
-                  "
-                  :selected="isFilterSelected(facet, option)"
-                  class="filters__item"
-                  @change="() => selectFilter(facet, option)"
-                />
-              </div>
-            </template>
-          </div>
-        </div>
-        <SfAccordion class="filters smartphone-only">
-          <div v-for="(facet, i) in facets" :key="i">
-            <SfAccordionItem
-              :key="`filter-title-${facet.id}`"
-              :header="facet.label"
-              class="filters__accordion-item"
-            >
-              <SfFilter
-                v-for="option in facet.options"
-                :key="`${facet.id}-${option.id}`"
-                :label="option.label"
-                :selected="isFilterSelected(facet, option)"
-                class="filters__item"
-                @change="() => selectFilter(facet, option)"
-              />
-            </SfAccordionItem>
-          </div>
-        </SfAccordion>
-
-        <template #content-bottom>
-          <div class="filters__buttons">
-            <SfButton class="sf-button--full-width" @click="applyFilters">{{
-              $t('Done')
-            }}</SfButton>
-            <SfButton
-              class="sf-button--full-width filters__button-clear"
-              @click="clearFilters"
-              >{{ $t('Clear all') }}</SfButton
-            >
-          </div>
-        </template>
-      </SfSidebar>
+      <CategoryFilterSideBar :facetsList="result" />
     </LazyHydrate>
   </div>
 </template>
 
 <script>
 import {
-  SfSidebar,
   SfButton,
   SfList,
-  SfIcon,
-  SfHeading,
   SfMenuItem,
-  SfFilter,
   SfProductCard,
+  SfHeading,
   SfProductCardHorizontal,
   SfPagination,
   SfAccordion,
   SfSelect,
-  SfBreadcrumbs,
-  SfCheckbox,
-  SfLoader,
-  SfColor,
   SfProperty,
+  SfBreadcrumbs,
+  SfLoader,
   SfImage
 } from '@storefront-ui/vue';
 import { ref, computed, onMounted } from '@nuxtjs/composition-api';
@@ -441,7 +273,6 @@ export default {
   transition: 'fade',
   setup(props, { root }) {
     const th = useUiHelpers();
-    const selectedFilters = ref([]);
     const generic = ref('');
 
     const { addTags } = useCache();
@@ -453,18 +284,13 @@ export default {
       isInWishlist
     } = useWishlist();
     const { result, search, loading } = useFacet();
-    const { params, query } = root.$router.history.current;
+    const { params } = root.$router.history.current;
 
     const products = computed(() => facetGetters.getProducts(result.value));
     const categoryTree = computed(() =>
       facetGetters.getCategoryTree(result.value)
     );
-    const sortBy = computed(() =>
-      facetGetters.getSortOptions({ input: { sort: query?.sort } } || '')
-    );
-    const facets = computed(() =>
-      facetGetters.getGrouped(result.value, ['color', 'size'])
-    );
+
     const pagination = computed(() => facetGetters.getPagination(result.value));
     const showProducts = computed(
       () => !loading.value && products.value?.length > 0
@@ -518,50 +344,9 @@ export default {
       ]);
     });
 
-    const { changeFilters, isFacetColor } = useUiHelpers();
-    const { toggleFilterSidebar } = useUiState();
     onMounted(() => {
       root.$scrollTo(root.$el, 2000);
-      selectedFilters.value = th.facetsFromUrlToFilter();
     });
-
-    const isFilterSelected = (facet, option) => {
-      return selectedFilters.value.some(
-        (filter) => String(filter.id) === String(option.value)
-      );
-    };
-
-    const selectFilter = (facet, option) => {
-      const alreadySelectedIndex = selectedFilters.value.findIndex(
-        (filter) => String(filter.id) === String(option.value)
-      );
-
-      if (alreadySelectedIndex === -1) {
-        selectedFilters.value.push({
-          filterName: facet.label,
-          label: option.label,
-          id: option.value
-        });
-
-        return;
-      }
-
-      selectedFilters.value.splice(alreadySelectedIndex, 1);
-    };
-
-    const clearFilters = () => {
-      toggleFilterSidebar();
-      selectedFilters.value = [];
-      changeFilters(selectedFilters.value);
-    };
-
-    const applyFilters = () => {
-      toggleFilterSidebar();
-      changeFilters(selectedFilters.value);
-    };
-
-    const facetHasMoreThanOneOption = (facet) =>
-      facet?.options?.length > 1 || false;
 
     return {
       ...uiState,
@@ -574,43 +359,30 @@ export default {
       loading,
       productGetters,
       pagination,
-      sortBy,
-      facets,
       breadcrumbs,
-      applyFilters,
       addItemToWishlist,
       removeItemFromWishList,
       addItemToCart,
       isInWishlist,
       isInCart,
-      isFacetColor,
-      selectFilter,
-      isFilterSelected,
-      selectedFilters,
-      clearFilters,
-      facetHasMoreThanOneOption,
       showProducts,
+      result,
       currentCategoryNameForAccordion
     };
   },
   components: {
+    SfSelect,
+    SfProperty,
     SfButton,
-    SfSidebar,
-    SfIcon,
     SfList,
-    SfFilter,
     SfProductCard,
     SfProductCardHorizontal,
     SfPagination,
-    SfCheckbox,
     SfMenuItem,
+    SfHeading,
     SfAccordion,
-    SfSelect,
     SfBreadcrumbs,
     SfLoader,
-    SfColor,
-    SfHeading,
-    SfProperty,
     LazyHydrate,
     SfImage
   }
@@ -637,137 +409,7 @@ export default {
   margin: var(--spacer-base) auto var(--spacer-lg);
   text-transform: capitalize;
 }
-.navbar {
-  position: relative;
-  display: flex;
-  border: 1px solid var(--c-light);
-  border-width: 0 0 1px 0;
-  @include for-desktop {
-    border-width: 1px 0 1px 0;
-  }
-  &.section {
-    padding: var(--spacer-sm);
-    @include for-desktop {
-      padding: 0;
-    }
-  }
-  &__aside,
-  &__main {
-    display: flex;
-    align-items: center;
-    padding: var(--spacer-sm) 0;
-  }
-  &__aside {
-    flex: 0 0 15%;
-    padding: var(--spacer-sm) var(--spacer-sm);
-    border: 1px solid var(--c-light);
-    border-width: 0 1px 0 0;
-  }
-  &__main {
-    flex: 1;
-    padding: 0;
-    justify-content: space-between;
-    @include for-desktop {
-      padding: var(--spacer-xs) var(--spacer-xl);
-    }
-  }
-  &__title {
-    --heading-title-font-weight: var(--font-weight--semibold);
-    --heading-title-font-size: var(--font-size--xl);
-  }
-  &__filters-icon {
-    margin: 0 0 0 var(--spacer-xs);
-    order: 1;
-    @include for-desktop {
-      margin: 0 var(--spacer-xs) 0 0;
-      order: 0;
-    }
-  }
-  &__filters-button {
-    display: flex;
-    align-items: center;
-    --button-font-size: var(--font-size--base);
-    --button-text-decoration: none;
-    --button-color: var(--c-link);
-    --button-font-weight: var(--font-weight--normal);
-    @include for-mobile {
-      --button-font-weight: var(--font-weight--medium);
-      order: 2;
-    }
-    svg {
-      fill: var(--c-text-muted);
-      transition: fill 150ms ease;
-    }
-    &:hover {
-      svg {
-        fill: var(--c-primary);
-      }
-    }
-  }
-  &__label {
-    font-family: var(--font-family--secondary);
-    font-weight: var(--font-weight--normal);
-    color: var(--c-text-muted);
-    @include for-desktop {
-      color: var(--c-link);
-      margin: 0 var(--spacer-2xs) 0 0;
-    }
-  }
-  &__select {
-    --select-width: 220px;
-    --select-padding: 0;
-    --select-height: auto;
-    --select-selected-padding: 0 var(--spacer-lg) 0 var(--spacer-2xs);
-    --select-margin: 0;
-    --select-option-font-size: var(--font-size-sm);
-    --select-error-message-height: 0;
-    ::v-deep .sf-select__dropdown {
-      font-size: var(--font-size-sm);
-      font-family: var(--font-family--secondary);
-      font-weight: var(--font-weight--light);
-      margin: 0;
-    }
-    ::v-deep .sf-select__placeholder {
-      --select-option-font-size: var(--font-size-sm);
-    }
-  }
-  &__sort {
-    display: flex;
-    align-items: center;
-    margin: 0 auto 0 var(--spacer-2xl);
-  }
-  &__counter {
-    font-family: var(--font-family--secondary);
-    order: 1;
-    @include for-desktop {
-      margin: auto 0 auto auto;
-      order: 0;
-    }
-  }
-  &__view {
-    display: flex;
-    align-items: center;
-    order: 0;
-    @include for-desktop {
-      margin: 0 0 0 var(--spacer-2xl);
-      order: 0;
-    }
-    &-icon {
-      cursor: pointer;
-      margin: 0 var(--spacer-base) 0 0;
-      &:last-child {
-        margin: 0;
-      }
-    }
-    &-label {
-      margin: 0 var(--spacer-sm) 0 0;
-      font: var(--font-weight--normal) var(--font-size--base) / 1.6
-        var(--font-family--secondary);
-      text-decoration: none;
-      color: var(--c-link);
-    }
-  }
-}
+
 .sort-by {
   flex: unset;
   width: 11.875rem;
@@ -775,19 +417,12 @@ export default {
 .main {
   display: flex;
 }
+
 .sidebar {
   flex: 0 0 15%;
   padding: var(--spacer-sm);
   border: 1px solid var(--c-light);
   border-width: 0 1px 0 0;
-}
-.sidebar-filters {
-  --sidebar-title-display: none;
-  --sidebar-top-padding: 0;
-  @include for-desktop {
-    --sidebar-content-padding: 0 var(--spacer-xl);
-    --sidebar-bottom-padding: 0 var(--spacer-xl);
-  }
 }
 .list {
   --menu-item-font-size: var(--font-size--sm);
@@ -876,63 +511,7 @@ export default {
 ::v-deep .sf-sidebar__aside {
   --sidebar-z-index: 3;
 }
-.filters {
-  &__title {
-    --heading-title-font-size: var(--font-size--xl);
-    margin: var(--spacer-xl) 0 var(--spacer-base) 0;
-    &:first-child {
-      margin: calc(var(--spacer-xl) + var(--spacer-base)) 0 var(--spacer-xs) 0;
-    }
-  }
-  &__colors {
-    display: flex;
-  }
-  &__color {
-    margin: var(--spacer-xs) var(--spacer-xs) var(--spacer-xs) 0;
-  }
-  &__chosen {
-    color: var(--c-text-muted);
-    font-weight: var(--font-weight--normal);
-    font-family: var(--font-family--secondary);
-    position: absolute;
-    right: var(--spacer-xl);
-  }
-  &__item {
-    --radio-container-padding: 0 var(--spacer-sm) 0 var(--spacer-xl);
-    --radio-background: transparent;
-    --filter-label-color: var(--c-secondary-variant);
-    --filter-count-color: var(--c-secondary-variant);
-    --checkbox-padding: 0 var(--spacer-sm) 0 var(--spacer-xl);
-    padding: var(--spacer-sm) 0;
-    border-bottom: 1px solid var(--c-light);
-    &:last-child {
-      border-bottom: 0;
-    }
-    @include for-desktop {
-      --checkbox-padding: 0;
-      margin: var(--spacer-sm) 0;
-      border: 0;
-      padding: 0;
-    }
-  }
-  &__accordion-item {
-    --accordion-item-content-padding: 0;
-    position: relative;
-    left: 50%;
-    right: 50%;
-    margin-left: -50vw;
-    margin-right: -50vw;
-    width: 100vw;
-  }
-  &__buttons {
-    margin: var(--spacer-sm) 0;
-  }
-  &__button-clear {
-    --button-background: var(--c-light);
-    --button-color: var(--c-dark-variant);
-    margin: var(--spacer-xs) 0 0 0;
-  }
-}
+
 .before-results {
   box-sizing: border-box;
   padding: var(--spacer-lg) var(--spacer-sm) var(--spacer-2xl);
